Rename JS source files in CamelCase, as in ui-directory
