--- src/components/patron-request/index.js
@@ -1,1 +1,0 @@
-export { default } from './view-patron-request';
